+++ geotrek/feedback/management/commands/sync_suricate.py
@@ -51,7 +51,7 @@ class Command(BaseCommand):
 
     def handle(self, *args, **options):
         verbosity = options['verbosity']
-        if settings.SURICATE_MANAGEMENT_ENABLED or settings.SURICATE_WORKFLOW_ENABLED:
+        if settings.SURICATE_WORKFLOW_ENABLED:
             parser = SuricateParser()
             has_no_params = not (options["statuses"] | options["activities"] | options["test"])
             report = options["report"]
@@ -68,4 +68,4 @@ class Command(BaseCommand):
                 if has_no_params:
                     parser.get_alerts(verbosity=verbosity, should_notify=not (no_notification))
         else:
-            logger.error("To use this command, please activate setting SURICATE_MANAGEMENT_ENABLED or SURICATE_WORKFLOW_ENABLED.")
+            logger.error("To use this command, please activate setting SURICATE_WORKFLOW_ENABLED.")