+++ geotrek/feedback/models.py
@@ -90,6 +90,12 @@ class PendingSuricateAPIRequest(models.Model):
     def remove_sync_error_flag_on_report(self, external_uuid):
         report = Report.objects.filter(external_uuid=external_uuid)
         report.update(sync_errors=F('sync_errors') - 1)
+        report = report.first()
+        if self.endpoint == "wsUpdateStatus" and json.loads(self.params).get('statut', '') == 'waiting' and report.status.identifier == 'filed':
+            # "waiting" status from suricate API does not override internal statuses on sync_suricate
+            # therefore, we need to manually set report status to waiting here if this is the call that failed
+            report.status = ReportStatus.objects.get(identifier='waiting')
+            report.save(update_fields=['status'])
 
     def save(self, *args, **kwargs):
         # Set sync_errors flag on report
@@ -189,7 +195,7 @@ class Report(GeotrekMapEntityMixin, PicturesMixin, TimeStampedModelMixin, NoDele
         ordering = ["-date_insert"]
 
     def __str__(self):
-        if (settings.SURICATE_WORKFLOW_ENABLED or settings.SURICATE_MANAGEMENT_ENABLED) and self.eid:
+        if settings.SURICATE_WORKFLOW_ENABLED and self.eid:
             return f"{_('Report')} {self.eid}"
         else:
             return f"{_('Report')} {self.pk}"
@@ -251,23 +257,13 @@ class Report(GeotrekMapEntityMixin, PicturesMixin, TimeStampedModelMixin, NoDele
             self.get_suricate_messenger().post_report(self)
         super().save(*args, **kwargs)  # Report updates should do nothing more
 
-    def save_suricate_management_mode(self, *args, **kwargs):
-        """Save method for Suricate Management mode"""
-        if not self.pk:  # This is a new report
-            if self.external_uuid is None:  # This new report comes from Rando or Admin : let Suricate handle it first, don't even save it
-                self.get_suricate_messenger().post_report(self)
-            else:  # This new report comes from Suricate : save
-                super().save(*args, **kwargs)
-        else:  # Report updates should do nothing more
-            super().save(*args, **kwargs)
-
     def save_suricate_workflow_mode(self, *args, **kwargs):
         """Save method for Suricate Management mode"""
         if not self.pk:  # This is a new report
             if self.external_uuid is None:  # This new report comes from Rando or Admin : let Suricate handle it first, don't even save it
                 self.get_suricate_messenger().post_report(self)
             else:  # This new report comes from Suricate : assign workflow manager if needed and save
-                if self.status.identifier in ['filed']:
+                if self.status.identifier in ['filed'] and not settings.SURICATE_WORKFLOW_SETTINGS.get("SKIP_MANAGER_MODERATION"):
                     self.assigned_user = WorkflowManager.objects.first().user
                 super().save(*args, **kwargs)
         else:  # Report updates should do nothing more
@@ -354,12 +350,10 @@ class Report(GeotrekMapEntityMixin, PicturesMixin, TimeStampedModelMixin, NoDele
                 self.get_suricate_messenger().message_sentinel(self.formatted_external_uuid, message_sentinel)
 
     def save(self, *args, **kwargs):
-        if not settings.SURICATE_REPORT_ENABLED and not settings.SURICATE_MANAGEMENT_ENABLED and not settings.SURICATE_WORKFLOW_ENABLED:
+        if not settings.SURICATE_REPORT_ENABLED and not settings.SURICATE_WORKFLOW_ENABLED:
             self.save_no_suricate(*args, **kwargs)  # No Suricate Mode
-        elif settings.SURICATE_REPORT_ENABLED and not settings.SURICATE_MANAGEMENT_ENABLED and not settings.SURICATE_WORKFLOW_ENABLED:
+        elif settings.SURICATE_REPORT_ENABLED and not settings.SURICATE_WORKFLOW_ENABLED:
             self.save_suricate_report_mode(*args, **kwargs)  # Suricate Report Mode
-        elif settings.SURICATE_MANAGEMENT_ENABLED and not settings.SURICATE_WORKFLOW_ENABLED:
-            self.save_suricate_management_mode(*args, **kwargs)  # Suricate Management Mode
         elif settings.SURICATE_WORKFLOW_ENABLED:
             self.save_suricate_workflow_mode(*args, **kwargs)  # Suricate Workflow Mode
 
@@ -384,7 +378,7 @@ class Report(GeotrekMapEntityMixin, PicturesMixin, TimeStampedModelMixin, NoDele
 
     @property
     def eid_verbose_name(self):
-        return _("Tag") if (settings.SURICATE_WORKFLOW_ENABLED or settings.SURICATE_MANAGEMENT_ENABLED) else _("Label")
+        return _("Tag") if settings.SURICATE_WORKFLOW_ENABLED else _("Label")
 
     def distance(self, to_cls):
         """Distance to associate this report to another class"""